fix: bug save product cart fixed

+++ src/shopping_cart/application/shopping_cart_product_deleter.py
@@ -6,5 +6,13 @@ class ShoppingCartProductDeleter:
         self._cart_repository = shopping_cart_repository
 
     def delete(self, uuid_cart: str, uuid_product: str) -> str:
-        self._cart_repository.delete_product_by_id(uuid_cart, uuid_product)
+        product = self._cart_repository.get_product_by_id(uuid_cart, uuid_product)
+
+        if product is not None:
+            if product.quantity == 1:
+                self._cart_repository.delete_product_by_id(uuid_cart, uuid_product)
+            else:
+                product.quantity -= 1
+                self._cart_repository.save_product(uuid_cart, product)
+        
         return uuid_product
